Provide meta/run.yam and README for Java and NFS
With this commit we provide README.md for all three OpenJDK Java
recipes that we currently support and, more importantly, a meta/run.yaml
for each of them. Having meta/run.yaml as part of the package it now makes
it extremely simple to run the package without knowing its structure.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>