Update and document mysql package
When testing the mysql package I've noticed that there are
some files missing in package, namely:

* /usr/data/performance_schema/ (whole directory)
* /usr/data/ibdata1

After some debugging it turned out that the two missing things
normally appear on my Ubuntu16 machine, but I've failed to
understand why. So for now, since MySQL package is not of high
priority, I've simply grabbed the missing things and added them
to the recipe.

Also, with this commit we add documentation and meta/run.yaml for
the MySQL package.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>