Update developer documentation
Developer documentation i.e. documentation about how to develop new
recipe has not been complete. With this commit, we fill in the new
environment variables description that were not described before and
notify developer that she can connect to the container and manually
experiment there. Also, we add a section that describes how to add
package README easily.

Since we've just discovered that adding `--privileged` flag to the
container run command decreases unikernel compose time significantly
as it enables KVM, we document this as well.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>