Provide sample recipe that can be used for debugging
When implementing enhancments of capstan-packages.py script, we
need to debug them on some recipe. But running actual recipe takes
a long time, usually two minutes or more, since it compiles some
source code. With this commit we provide a sample recipe that instantly
produces a valid capstan package to cover such cases.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>