Update apache.spark-2.1.1 recipe to version 0.2
Old recipe for Spark didn't play well anymore due to some changes
on dependent repositories. Most notably, the osvProcessBuilder repository
is now configured as a Maven project so manually calling `make` program
is no longer needed. Also, since it's a Maven project, it puts the
result in target/ directory and not in some 5-level deeply nested direcotry
like in case we call `make` manually.

With this commit we mostly simplify the recipe so that it's now nicely
readable. Also, since we've modified the recipe, we increase packaging
version from 0.1 to 0.2.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>