Include decomposePar.so and reconstructPar.so in openfoam.core-2.4.0
We're intentionally excluding ROOTFS/usr/bin/** from OpenFOAM package
since the bin directory contains solvers that we ship in separate packages.
But it turns out that there are also decomposePar.so and reconstructPar.so
files in this same directory that should be included in core package. With
this commit we therefore manually include them into the package.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>