Follow upstream changes of scripts/build
Until now we had to manually patch the scripts/build script in order
to collect build result into a directory instead into unikernel disk.
But recently the changes got upstreamed:

https://github.com/cloudius-systems/osv/commit/279effd05f2ee32d4e7efa44e607d55ecf3202e6

Since the upstreamed changes are slightly different, we have to adjust
all the recipes. With this commit we first discard the two patches that
are no longer needed and then we update the first two recipes: osv.bootstrap
and openjdk7 to demonstrate how all recipes should be adjusted.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>